move test to another place
