+++ Makefile.am
@@ -519,7 +519,6 @@ libarchive_test_SOURCES= \
 	libarchive/test/test_read_format_xar.c \
 	libarchive/test/test_read_format_zip.c \
 	libarchive/test/test_read_format_zip_7075_utf8_paths.c \
-	libarchive/test/test_read_format_zip_7z_lzma.c \
 	libarchive/test/test_read_format_zip_comment_stored.c \
 	libarchive/test/test_read_format_zip_encryption_data.c \
 	libarchive/test/test_read_format_zip_encryption_partially.c \
