+++ libarchive/test/CMakeLists.txt
@@ -170,7 +170,6 @@ IF(ENABLE_TEST)
     test_read_format_xar.c
     test_read_format_zip.c
     test_read_format_zip_7075_utf8_paths.c
-    test_read_format_zip_7z_lzma.c
     test_read_format_zip_comment_stored.c
     test_read_format_zip_encryption_data.c
     test_read_format_zip_encryption_header.c
